Major bug fix, more refactoring
The last update introduced a bug, the check that psi is real was wrong, so it was acting like psi was real when it wasn't. This has been fixed. There's also another ~25% boost in performance by moving essentially everything into the H_G function, making better use of the cache by avoiding looping through b2 a second time.